# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Scootware Forum API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}/avatar:
    post:
      operationId: uploadAvatar
      tags: [users]
      summary: Upload user avatar
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
      responses:
        "200":
          description: Avatar uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}/posts:
    get:
      operationId: getUserProfilePosts
      tags: [users]
      summary: Get posts on user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfilePost"
    post:
      operationId: createProfilePost
      tags: [users]
      summary: Post on a user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfilePostRequest"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfilePost"
  /users/me/invites:
    get:
      operationId: getMyInvites
      tags: [users]
      summary: Get current user's issued invites and their usage
      responses:
        "200":
          description: User invites data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInvitesResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/request-invite:
    post:
      operationId: requestInviteAuthenticated
      tags: [users]
      summary: Request an invite (authenticated users in invite-only mode)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestInviteAuthenticatedRequest"
      responses:
        "200":
          description: Invite request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users/{userId}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        upgradeType:
          type: string
          nullable: true
          enum: [
            BODYCAM_PREMIUM, BODYCAM_LIFETIME,
            RUST_PREMIUM, RUST_LIFETIME,
            DAYZ_PREMIUM, DAYZ_LIFETIME,
            TARKOV_PREMIUM, TARKOV_LIFETIME,
            SPOOFER_PREMIUM, SPOOFER_LIFETIME
          ]
        upgradeExpiresAt:
          type: string
          format: date-time
          nullable: true
        productAccess:
          type: array
          items:
            $ref: "#/components/schemas/UserProductAccess"
          description: List of products the user has access to via invites or purchases
        avatarUrl:
          type: string
          nullable: true
        isBanned:
          type: boolean
        isEmailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        postCount:
          type: integer
      required: [id, username, email, role, isBanned, isEmailVerified, createdAt, postCount, productAccess]
    UserProductAccess:
      type: object
      properties:
        productId:
          type: string
        expiresAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        recentPosts:
          type: array
          items:
            $ref: "#/components/schemas/ProfilePost"
      required: [user, recentPosts]
    ProfilePost:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        authorId:
          type: integer
        authorUsername:
          type: string
        authorAvatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, content, authorId, authorUsername, createdAt]
    CreateProfilePostRequest:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
      required: [content]
...
    AdminUpdateUserRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        role:
          type: string
          enum: [user, admin]
        upgradeType:
          type: string
          nullable: true
          enum: [
            BODYCAM_PREMIUM, BODYCAM_LIFETIME,
            RUST_PREMIUM, RUST_LIFETIME,
            DAYZ_PREMIUM, DAYZ_LIFETIME,
            TARKOV_PREMIUM, TARKOV_LIFETIME,
            SPOOFER_PREMIUM, SPOOFER_LIFETIME
          ]
        upgradeExpiresAt:
          type: string
          format: date-time
          nullable: true
...
    UserInvitesResponse:
      type: object
      properties:
        invites:
          type: array
          items:
            $ref: "#/components/schemas/InviteCodeWithUsageInfo"
      required: [invites]
    InviteCodeWithUsageInfo:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        productId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
        isUsed:
          type: boolean
        usedBy:
          type: integer
          nullable: true
        usedByUsername:
          type: string
          nullable: true
        usedByAvatarUrl:
          type: string
          nullable: true
        usedAt:
          type: string
          format: date-time
          nullable: true
        isBanned:
          type: boolean
      required: [id, code, createdAt, isUsed, isBanned]
    RequestInviteAuthenticatedRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 1000
          description: Optional reason for requesting an invite